+++ CHANGELOG.md
@@ -4,13 +4,16 @@
 
 
 ### FEATURES
-
 - Continous Integration (CI) end-to-end (e2e) testing ([#32])
 - Add support for streamlining releases ([#507])
+
 - [relayer-cli]
   - Implement command to query the channels associated with a connection ([#505])
   - JSON output for queries and txs ([#500])
   - Added 'required' annotation for CLIs queries & txs; better error display ([#555])
+  - Implement commands for channel close init and confirm ([#538])
+  - Implement command to perform the handshake for a new channel ([#557])
+  
 - [relayer]
   - Added retry mechanism, restructured relayer ([#519])
 
@@ -22,7 +25,7 @@
 - [relayer-cli]
   - Replace `ChannelConfig` in `Channel::new` ([#511])
   - Add `packet-send` CLI ([#470])
-  - UX improvements for relayer txs ([#536, #540])
+  - UX improvements for relayer txs ([#536, #540, #554])
 
 - [relayer]
   - Performance improvements ([#514], [#537])
@@ -59,9 +62,11 @@
 [#535]: https://github.com/informalsystems/ibc-rs/issues/535
 [#536]: https://github.com/informalsystems/ibc-rs/issues/536
 [#537]: https://github.com/informalsystems/ibc-rs/issues/537
+[#538]: https://github.com/informalsystems/ibc-rs/issues/538
 [#540]: https://github.com/informalsystems/ibc-rs/issues/540
 [#555]: https://github.com/informalsystems/ibc-rs/issues/555
-
+[#554]: https://github.com/informalsystems/ibc-rs/issues/554
+[#557]: https://github.com/informalsystems/ibc-rs/issues/557
 
 ## v0.0.6
 *December 23, 2020*